add star func and getAvgStar func

+++ src/main/java/com/team5/c4quanlyphongsach/model/Book.java
@@ -42,5 +42,5 @@ public class Book {
     @ManyToOne
     private Customer customer;
 
-    private Long rate;
+    private Float rate;
 }
